Made menu buttons clickable.
