+++ core/context/GraphicsContext.h
@@ -4,7 +4,7 @@ class GraphicsContext{
 	public:
 		int context = 0;
 		GraphicsCamera *activeCamera = NULL;
-
+		
 		virtual void destroy(){}
 		virtual void init(){}
 		virtual int exec(GLFWwindow* window){
