+++ core/graphics/GraphicsTexture.h
