+++ core/graphics/graphics.h
@@ -1,0 +1,6 @@
+#include "./GraphicsShader.h"
+#include "./GraphicsTexture.h"
+#include "./GraphicsCamera.h"
+
+#include "./GraphicsObject.h"
+
