+++ core/peripherals/peripheralCallbacks.h
@@ -1,127 +1,3 @@
-typedef struct guiEngineKeyboard{
-	bool key_a = false;
-	bool key_b = false;
-	bool key_c = false;
-	bool key_d = false;
-	bool key_e = false;
-	bool key_f = false;
-	bool key_g = false;
-	bool key_h = false;
-	bool key_i = false;
-	bool key_j = false;
-	bool key_k = false;
-	bool key_l = false;
-	bool key_m = false;
-	bool key_n = false;
-	bool key_o = false;
-	bool key_p = false;
-	bool key_q = false;
-	bool key_r = false;
-	bool key_s = false;
-	bool key_t = false;
-	bool key_u = false;
-	bool key_v = false;
-	bool key_w = false;
-	bool key_x = false;
-	bool key_y = false;
-	bool key_z = false;
-
-	bool key_A = false;
-	bool key_B = false;
-	bool key_C = false;
-	bool key_D = false;
-	bool key_E = false;
-	bool key_F = false;
-	bool key_G = false;
-	bool key_H = false;
-	bool key_I = false;
-	bool key_J = false;
-	bool key_K = false;
-	bool key_L = false;
-	bool key_M = false;
-	bool key_N = false;
-	bool key_O = false;
-	bool key_P = false;
-	bool key_Q = false;
-	bool key_R = false;
-	bool key_S = false;
-	bool key_T = false;
-	bool key_U = false;
-	bool key_V = false;
-	bool key_W = false;
-	bool key_X = false;
-	bool key_Y = false;
-	bool key_Z = false;
-	
-	bool key_0 = false;
-	bool key_1 = false;
-	bool key_2 = false;
-	bool key_3 = false;
-	bool key_4 = false;
-	bool key_5 = false;
-	bool key_6 = false;
-	bool key_7 = false;
-	bool key_8 = false;
-	bool key_9 = false;
-	bool key_exclame = false;
-	bool key_at = false;
-	bool key_pound = false;
-	bool key_money = false;
-	bool key_percent = false;
-	bool key_carrot = false;
-	bool key_amp = false;
-	bool key_star = false;
-	bool key_popen = false;
-	bool key_pclose = false;
-	bool key_minus = false;
-	bool key_underscore = false;
-	bool key_equals = false;
-	bool key_plus = false;
-	bool key_back = false;
-	bool key_tilde = false;
-	bool key_backtick = false;
-
-	bool key_enter = false;
-	bool key_tab = false;
-	bool key_esc = false;
-	bool key_space = false;
-	
-	bool key_up = false;
-	bool key_down = false;
-	bool key_left = false;
-	bool key_right = false;
-	bool key_leftshift = false;
-	bool key_rightshift = false;
-	bool key_leftctrl = false;
-	bool key_rightctrl = false;
-	bool key_leftalt = false;
-	bool key_home = false;
-	bool key_end = false;
-	bool key_pageup = false;
-	bool key_pagedown = false;
-	bool key_insert = false;
-
-}gui_engine_keyboard_t;
-
-typedef struct guiEngineMouse{
-	double mouseX = 0;
-	double mouseY = 0;
-	bool click_left = false;
-	bool click_right = false;
-	bool click_middle = false;
-	bool scroll_up = false;
-	bool scroll_down = false;
-}gui_engine_mouse_t;
-
-struct GuiEngineStruct{
-        //string windowTitle = "Gui Engine";
-	int displayX = 800;
-	int displayY = 600;	
-	gui_engine_keyboard_t keyboard;
-	gui_engine_mouse_t mouse;
-
-}gui_engine_global;
-
 void updateKeyboardSpecial(bool keyDown, int key, int mouseX, int mouseY){
 	gui_engine_global.mouse.mouseX = mouseX;
 	gui_engine_global.mouse.mouseY = mouseY;
