+++ core/peripherals/peripherals.h
@@ -1,0 +1,1 @@
+#include "./peripheralCallbacks.h"
