+++ core/structs/guiEngine.struct.h
@@ -1,0 +1,8 @@
+struct GuiEngineStruct{
+        //string windowTitle = "Gui Engine";
+	int displayX = 800;
+	int displayY = 600;	
+	gui_engine_keyboard_t keyboard;
+	gui_engine_mouse_t mouse;
+
+}gui_engine_global;
