+++ core/structs/guiEngineMouse.struct.h
@@ -1,0 +1,9 @@
+typedef struct guiEngineMouse{
+	double mouseX = 0;
+	double mouseY = 0;
+	bool click_left = false;
+	bool click_right = false;
+	bool click_middle = false;
+	bool scroll_up = false;
+	bool scroll_down = false;
+}gui_engine_mouse_t;
