+++ core/structs/structs.h
@@ -1,0 +1,3 @@
+#include "./guiEngineKeyboard.struct.h"
+#include "./guiEngineMouse.struct.h"
+#include "./guiEngine.struct.h"
